refactor(utils): extract shared formatting utilities to src/utils/format.ts
Consolidate duplicated helper functions and constants spread across
ScheduleDetailPage, WebhookDetailPage, PostActionDetailPage,
PostActionsListPage, SchedulesListPage, and WebhooksListPage into a
single module.

- formatDateTime, formatDuration, formatRelativeTime, sanitizeRunError
  were defined 2-3 times identically across detail and list pages
- METHOD_COLORS and AUTH_LABELS were duplicated between PostActionDetailPage
  and PostActionsListPage with an inconsistency: the list page used
  abbreviated labels ('Bearer', 'Basic') while the detail page used full
  labels ('Bearer Token', 'Basic Auth') — now unified to the full form
- runStatusStyles was duplicated across all three detail pages and is now
  exported as RUN_STATUS_STYLES
- PostActionDetailPage had a bug in its local formatDuration: the hours
  branch was missing, so runs longer than 60 minutes displayed as e.g.
  '90m 30s' instead of '1h 30m' — fixed by using the shared implementation
